Ensure all documents can be retrieved
Last time I had run the project it was working propertly when fetching all documents, but we had to make these changes to ensure that fetching all documents would work again.

The changes in this commit may not be as robust as needed, but that will be determined later, as basic functionality appears to be working well enough that we can move forward.