that's the FACADE
The docstring was updated in the verba_manager to descibe the facade design pattern that encapsulates methods and their managers.

Please note that the only function defintion that maintained @bool and @parameter and @returns is 'check_if_document_exits'

I am not sure what these are supposed to be used, but I assume that it's for documentation display.