Video Editing and Frame desc. Retrieval
With this commit we have now have the ability to use RAG to search for the frame a description of a frame and return structured data about the frame.

We have also included the ability to generate editing instructions using the VideoEditingGenerator.

Note that this version of the code sets the cashing to False on load and also sets the generator to VideoEditing on load.