important VideoEditingGenerator changes + Doc Type Filtering Retriever and Planner
let's begin with crucial working changes. The VideoEditingGenerator.py is confirmed to be working correctly and is the most important part of this commit.

The doc type filtering retriever has not been tested. There may be a more straight forward way to do this with the original retriever. I think you'd review the interfaces.py file to be sure, but I am not certain.

The planner generator does not work yet.