TranscriptionReader, Chunker and Generator
The API Gateway is currently responsible for the cleaning of the deepgram transcription JSON and doesn't include the full transcript under a single property. Since the documents needs a full text at the document level, the reader concatenate the utterances and joins them to create the full text.

The chunker creates a "chunk info" and utterance timing that is aligned with the chunked utterance.

in import_data:

1. It's iterating through the existing chunks of each document.
2. It's creating batches of chunks, where each batch has a maximum of 4000 tokens.
3. If adding a chunk would exceed 4000 tokens, it starts a new batch.

At the time of writing this I have not yet tested to see what it will do beyond 4000 tokens.

Generator:
The generator may be vulnerable to hallucination as it gets the structured data from the chunks, and many chunks are sent as context.

Note that only transcriptionSearchGenerator is called with chunk_info in generate_stream_answer

also note that 'transcriptionSearchGenerator ' isn't an idea name and will probably change later.

finally, note that some of the chanks to the chunk.py class might be unnecessary. The retrieve_chunks  function in the Verba Manager class has code that if it can't find matching chunk_info, it creates a fallback chunk_info using the attributes of the Chunk object.

I am not entirely sure how this works, but if the current changes in chunk.py are reverted, the chunk_info doesn't make it to the transcriptionSearchGenerator. The chunk_info is still available, which means there's some issue with the design here. This is tech debt.