Generate Chad Meme API endpoint
Added this new endpoint and a StoryJSONGenerator to take a story idea from plain text and use multiple GPT4 endpoint calls to build out a detailed JSON object describing the video editing layers which are finally parsed into a list of timeline_commands that can be executed by the kdenlive client.

This commit also includes some truncation for logging which was blowing up the terminal.

This commit also includes 'TranscriptionFinder' which doesn't really do anything yet. The idea for this component is that it will be a faster version of the existing TranscriptionSearchGenerator, which is unnecessarily slow and bloated compared to the v1 2023 protype that found video moments or deepgram transcription moments very fast using only a similarity search and no gpt4 calls or gpt JSON mode calling.

Also added "WindowRetrieveSimilarity" which is designed to simply use a similarity search rather than keyword hybrid. It seems to be significantly faster, but the chunk retrieval that wraps it slows down requests by a very large amount of time.