update StoryJSONGenerator
Now it creates a pydantic valid output in multiple steps. The context is passed as part of the prompt to inform the LLM API endpoint about available assets and how to write the story.

For example "Use only white backgrounds" or "keep the pacing fast. Usually 2.0 seconds per shot" or other instructions.

At the moment the verba would probably fail to generate good outputs if there's a lot of files in the context database. Later we'll filter these instructions to ensure that only instructions files specific to the project or user preferences are included.

This is exciting!